fixed property types #89

+++ Billbee.Api.Client/Model/DeletedImages.cs
@@ -8,7 +8,7 @@ namespace Billbee.Api.Client.Model
 {
     public class DeletedImages
     {
-        public List<int> Deleted { get; set; }
-        public List<int> NotFound { get; set; }
+        public List<long> Deleted { get; set; }
+        public List<long> NotFound { get; set; }
     }
 }
